backend/pyproject.toml
[project]
name = "plant-backend"
version = "0.1.0"
description = "Backend API for plant automation system"
authors = [{ name = "Codex" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi~=0.110",
    "uvicorn[standard]~=0.30",
    "sqlalchemy[asyncio]~=2.0",
    "aiosqlite~=0.20",
    "python-multipart~=0.0.9",
    "bcrypt==4.0.1",
    "pydantic[email]~=2.12",
    "pydantic-settings~=2.4",
    "alembic~=1.13",
    "asyncpg~=0.29",
    "python-jose[cryptography]~=3.3",
    "passlib[bcrypt]~=1.7",
    "redis~=5.0",
    "itsdangerous~=2.2",
    "jinja2~=3.1"
]

[project.optional-dependencies]
dev = [
    "pytest~=8.2",
    "httpx~=0.27",
    "pytest-asyncio~=0.23"
]
pi = [
    "requests~=2.32"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*", "clients*", "scripts*"]
